+++ idf_component_tools/sources/web_service.py
@@ -22,7 +22,7 @@ from ..archive_tools import ArchiveError, get_format_from_path, unpack_archive
 from ..constants import IDF_COMPONENT_REGISTRY_URL, IDF_COMPONENT_STORAGE_URL, UPDATE_SUGGESTION
 from ..errors import FetchingError
 from ..file_tools import copy_directory
-from ..hash_tools import validate_filtered_dir
+from ..hash_validator import validate_managed_component
 from ..messages import hint
 from ..registry.base_client import create_session
 from . import utils
@@ -275,7 +275,9 @@ class WebServiceSource(BaseSource):
             return download_path
 
         # Check if component is in the cache
-        if validate_filtered_dir(self.component_cache_path(component), component.component_hash):
+        if validate_managed_component(
+            self.component_cache_path(component), component.component_hash
+        ):
             copy_directory(self.component_cache_path(component), download_path)
             return download_path